Assembly Dump Test Programs:  * Added support for printing basic properties of fields; most of the properties    are the same, but we need to look into the DeclaringType property.
WinRT Test Stubs
 * Created two new WinRT Test Stubs for testing the library within an app package.

CxxReflect Library
 * Added new CxxReflectPlatform.il to build our own platform metadata file.  We
   still need to figure out a way to package it, and it is nowhere near complete

 * Added new WinRTIntegration functionality to use RoResolveNamespace and friends
   to find metadata files.

 * Created a new property sheet to share common vcxproj properties

 * Fixed a bunch of 64-bit integer conversion warnings.

+++ CxxReflect/PlatformWin32.cpp
@@ -40,11 +40,11 @@ namespace CxxReflect { namespace Detail {
         if (!CryptCreateHash(provider, CALG_SHA1, 0, 0, &hash))
             throw std::logic_error("wtf");
 
-        if (!CryptHashData(hash, first, last - first, 0))
+        if (!CryptHashData(hash, first, static_cast<DWORD>(last - first), 0))
             throw std::logic_error("wtf");
 
         Sha1Hash result = { 0 };
-        DWORD resultLength(result.size());
+        DWORD resultLength(static_cast<DWORD>(result.size()));
         if (!CryptGetHashParam(hash, HP_HASHVAL, result.data(), &resultLength, 0) || resultLength != 20)
             throw std::logic_error("wtf");
 
@@ -53,6 +53,7 @@ namespace CxxReflect { namespace Detail {
 
     bool FileExists(wchar_t const* path)
     {
+        
         return GetFileAttributes(path) != INVALID_FILE_ATTRIBUTES;
     }
